Removes incorrect comma.
The error becomes obvious when you see that the left hand side is a
listing of three items - 'a, b, and c' - and the right hand side is a
single item. This makes the entire sentence a listing of two items,
which should obviously be 'a and b' as opposed to 'a, and b'.